requirements.txt
albumentations==0.4.6
celery==5.2.7
Django==3.2.25
einops==0.8.2
Flask==2.2.5
Flask_Cors==5.0.0
flask_sqlalchemy==3.1.1
matplotlib==3.5.3
numpy==1.21.6
opencv_python==4.8.0.74
Pillow==9.4.0
PyWavelets==1.3.0
PyYAML==6.0.1
ruamel.base==1.0.0
scikit_image==0.19.3
scikit_learn==1.0.2
scipy==1.7.3
setuptools==65.6.3
timm==0.6.11
torch==1.13.1
torch_geometric==2.1.0
torchvision==0.14.1
tqdm==4.66.1
Werkzeug==2.2.2
